@@ -1,12 +1,12 @@
 # frozen_string_literal: true
 
-require 'spec_helper'
+require "spec_helper"
 
 #
 # To "downgrade" an admin to department_admin, it must first be removed from admins
 # create a new admin as department_admin.
 #
-describe 'Admin invite user', type: :system do
+describe "Admin invite user", type: :system do
   let!(:area) { create(:area) }
 
   let(:organization) { create(:organization) }
@@ -15,8 +15,8 @@ describe 'Admin invite user', type: :system do
   let(:user_manager) { create(:user, :user_manager, :confirmed, organization: organization) }
 
   let(:department_admin) do
-    user= create(:user, :confirmed, organization: organization)
-    user.roles << 'department_admin'
+    user = create(:user, :confirmed, organization: organization)
+    user.roles << "department_admin"
     user.areas << area
     user.save!
     user
@@ -28,15 +28,15 @@ describe 'Admin invite user', type: :system do
     visit decidim_admin.new_user_path
   end
 
-  it 'admin is able to create department admins' do
-    fill_the_form_for_department_admin('Cabello Loco', 'my@email.net')
+  it "admin is able to create department admins" do
+    fill_the_form_for_department_admin("Cabello Loco", "my@email.net")
     submit_form
     check_succeess
-    check_is_department_admin('my@email.net')
-    check_assigned_area(@user, area)
+    user = check_is_department_admin("my@email.net")
+    check_assigned_area(user, area)
   end
 
-  it 'admin is able to add department_admin role to existing user' do
+  it "admin is able to add department_admin role to existing user" do
     fill_the_form_for_department_admin(user_manager.name, user_manager.email)
     submit_form
     check_succeess
@@ -44,7 +44,7 @@ describe 'Admin invite user', type: :system do
     check_assigned_area(user_manager, area)
   end
 
-  context 'when departments are reorganized' do
+  context "when departments are reorganized" do
     let!(:new_area) { create(:area) }
 
     before do
@@ -61,8 +61,8 @@ describe 'Admin invite user', type: :system do
     end
   end
 
-  context 'when a department_admin is promoted to Admin' do
-    it 'no longer has the `department_admin` role' do
+  context "when a department_admin is promoted to Admin" do
+    it "no longer has the `department_admin` role" do
       department_admin
       fill_the_form_for_admin(department_admin.name, department_admin.email)
       submit_form
@@ -72,42 +72,43 @@ describe 'Admin invite user', type: :system do
   end
 
   def fill_the_form_for_department_admin(name, email, selected_area = area)
-    within 'form.new_user' do
+    within "form.new_user" do
       fill_in :user_name, with: name
       fill_in :user_email, with: email
-      find('#user_role').find("option[value='department_admin']").select_option
-      expect(page).to have_css('#user_area_id')
-      find('#user_area_id').find("option[value='#{selected_area.id}']").select_option
+      find("#user_role").find("option[value='department_admin']").select_option
+      expect(page).to have_css("#user_area_id")
+      find("#user_area_id").find("option[value='#{selected_area.id}']").select_option
     end
   end
 
-  def fill_the_form_for_admin(name, email, selected_area = area)
-    within 'form.new_user' do
+  def fill_the_form_for_admin(name, email, _selected_area = area)
+    within "form.new_user" do
       fill_in :user_name, with: name
       fill_in :user_email, with: email
-      find('#user_role').find("option[value='admin']").select_option
+      find("#user_role").find("option[value='admin']").select_option
     end
   end
 
   def submit_form
-    find('*[name=commit][type=submit]').click
+    find("*[name=commit][type=submit]").click
   end
 
   def check_succeess
-    expect(page).to have_content('Participant successfully invited.')
-    expect(page).to have_current_path '/admin/users'
+    expect(page).to have_content("Participant successfully invited.")
+    expect(page).to have_current_path "/admin/users"
   end
 
   def check_is_admin(email)
-    @user= Decidim::User.find_by_email(email).reload
-    expect(@user).to be_admin
-    expect(@user.roles).to_not include('department_admin')
-    expect(@user.areas).to be_empty
+    user = Decidim::User.find_by(email: email).reload
+    expect(user).to be_admin
+    expect(user.roles).not_to include("department_admin")
+    expect(user.areas).to be_empty
   end
 
   def check_is_department_admin(email)
-    @user= Decidim::User.find_by_email(email)
-    expect(@user.roles).to include('department_admin')
+    user = Decidim::User.find_by(email: email)
+    expect(user.roles).to include("department_admin")
+    user
   end
 
   def check_assigned_area(user, area)
