@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_dependency 'decidim/admin/users_controller'
+require_dependency "decidim/admin/users_controller"
 
 # Sort Admins by role and area
 ::Decidim::Admin::UsersController.class_eval do
@@ -13,8 +13,8 @@ require_dependency 'decidim/admin/users_controller'
   before_action :set_global_params
 
   def collection
-    users= original_collection
-    Decidim::Admin::UserAdminFilter.for(users,@query, @search_text,@role, current_locale)
+    users = original_collection
+    Decidim::Admin::UserAdminFilter.for(users, @query, @search_text, @role, current_locale)
   end
 
   # It is necessary to overwrite this method to correctly locate the user.
@@ -23,69 +23,70 @@ require_dependency 'decidim/admin/users_controller'
     @user ||= original_collection.find(params[:id])
   end
 
+  # rubocop: disable Metrics/CyclomaticComplexity
+  # rubocop: disable Metrics/PerceivedComplexity
   def show
     locale = params[:locale] || "ca"
     @user ||= original_collection.find(params[:id])
     @spaces = []
     @user.participatory_processes.each do |process|
-      if process.participatory_process_group then
-        if process.participatory_process_group&.name[locale] != '' then
-          type = process.participatory_process_group&.name[locale]
-        else
-          type = process.participatory_process_group&.name["ca"]
-        end
-      else
-        type = t("models.user.fields.process_type", scope: "decidim.admin")
-      end
+      type = if process.participatory_process_group
+               if process.participatory_process_group&.name&.[](locale) != ""
+                 process.participatory_process_group&.name&.[](locale)
+               else
+                 process.participatory_process_group&.name&.[]("ca")
+               end
+             else
+               t("models.user.fields.process_type", scope: "decidim.admin")
+             end
       process_title = process.title[locale]
-      if process_title == '' then
-        process_title = process.title["ca"]
-      end
+      process_title = process.title["ca"] if process_title == ""
 
-      @spaces.push({"title" => process_title,
-                    "type" => type,
-                    "area" => process.area.nil? ? "" : process.area&.name[locale],
-                    "created_at" => process.created_at,
-                    "private" => process.private_space?,
-                    "published" => process.published?})
+      @spaces.push("title" => process_title,
+                   "type" => type,
+                   "area" => process.area.nil? ? "" : process.area&.name&.[](locale),
+                   "created_at" => process.created_at,
+                   "private" => process.private_space?,
+                   "published" => process.published?)
     end
 
     @user.assemblies.each do |assembly|
-
-      if assembly.area && assembly.area.name then
-        area_name = assembly.area.name[locale]
-      else
-        area_name = ""
-      end
+      area_name = if assembly.area && assembly.area.name
+                    assembly.area.name[locale]
+                  else
+                    ""
+                  end
 
       assembly_title = assembly.title[locale]
-      if assembly_title == '' then
-        assembly_title = assembly.title["ca"]
-      end
+      assembly_title = assembly.title["ca"] if assembly_title == ""
 
-      @spaces.push({"title" => assembly_title,
-                    "type" => t("models.user.fields.assembly_type", scope: "decidim.admin"),
-                    "area" => area_name,
-                    "created_at" => assembly.created_at,
-                    "private" => assembly.private_space?,
-                    "published" => assembly.published?})
+      @spaces.push("title" => assembly_title,
+                   "type" => t("models.user.fields.assembly_type", scope: "decidim.admin"),
+                   "area" => area_name,
+                   "created_at" => assembly.created_at,
+                   "private" => assembly.private_space?,
+                   "published" => assembly.published?)
     end
 
-    if params[:sort_column] && (params[:sort_column] == 'published' || params[:sort_column] == 'private') && params[:sort_order] && params[:sort_order] == 'asc' then
-      @spaces.sort! { |x, y| x[params[:sort_column]] ? 0 : 1 <=> y[params[:sort_column]] ? 0 : 1 }
-    elsif params[:sort_column] && (params[:sort_column] == 'published' || params[:sort_column] == 'private') && params[:sort_order] && params[:sort_order] == 'desc' then
+    # rubocop: disable Style/NestedTernaryOperator
+    if params[:sort_column] && (params[:sort_column] == "published" || params[:sort_column] == "private") && params[:sort_order] && params[:sort_order] == "asc"
+      @spaces.sort! { |x, y| x[params[:sort_column]] ? 0 : (1 <=> y[params[:sort_column]] ? 0 : 1) }
+    elsif params[:sort_column] && (params[:sort_column] == "published" || params[:sort_column] == "private") && params[:sort_order] && params[:sort_order] == "desc"
       @spaces.sort! { |x, y| y[params[:sort_column]] ? 0 : 1 <=> x[params[:sort_column]] ? 0 : 1 }
-    elsif params[:sort_column] && params[:sort_order] && params[:sort_order] == 'asc' then
-      @spaces.sort! { |x, y| x[params[:sort_column]] <=> y[params[:sort_column]]}
-    elsif params[:sort_column] && params[:sort_order] && params[:sort_order] == 'desc' then
-      @spaces.sort! { |x, y| y[params[:sort_column]] <=> x[params[:sort_column]]}
+    elsif params[:sort_column] && params[:sort_order] && params[:sort_order] == "asc"
+      @spaces.sort! { |x, y| x[params[:sort_column]] <=> y[params[:sort_column]] }
+    elsif params[:sort_column] && params[:sort_order] && params[:sort_order] == "desc"
+      @spaces.sort! { |x, y| y[params[:sort_column]] <=> x[params[:sort_column]] }
     else
-      @spaces.sort! {|x, y| y["title"] <=> x["title"]}
+      @spaces.sort! { |x, y| y["title"] <=> x["title"] }
     end
+    # rubocop: enable Style/NestedTernaryOperator
   end
+  # rubocop: enable Metrics/CyclomaticComplexity
+  # rubocop: enable Metrics/PerceivedComplexity
 
-  def sort_spaces(column, order)
-    @spaces.sort! {|x, y| x[column] <=> y[column]}.reverse
+  def sort_spaces(column, _order)
+    @spaces.sort! { |x, y| x[column] <=> y[column] }.reverse
   end
 
   def set_global_params
