@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_dependency 'decidim/area'
+require_dependency "decidim/area"
 
 Decidim::Area.class_eval do
   before_destroy :abort_if_department_admins
@@ -11,13 +11,13 @@ Decidim::Area.class_eval do
                           association_foreign_key: :decidim_user_id,
                           validate: false
 
-  has_many  :participatory_process,
-            class_name: 'ParticipatoryProcess',
-            foreign_key: 'decidim_department_id'
+  has_many :participatory_process,
+           class_name: "ParticipatoryProcess",
+           foreign_key: "decidim_department_id"
 
-  has_many  :assemblies,
-            class_name: 'Assemblies',
-            foreign_key: 'decidim_area_id'
+  has_many :assemblies,
+           class_name: "Assemblies",
+           foreign_key: "decidim_area_id"
 
   def has_department_admins?
     Decidim::User.where(organization: organization).any? do |u|
