@@ -8,10 +8,10 @@ Decidim::DecidimFormHelper.class_eval do
   alias_method :original_areas_for_select, :areas_for_select
 
   def areas_for_select(organization)
-    author= current_user
+    author = current_user
 
     if author&.department_admin?
-      return author.areas if controller_path.split('/').include?('admin')
+      return author.areas if controller_path.split("/").include?("admin")
     end
 
     original_areas_for_select(organization)
