@@ -10,10 +10,10 @@ Decidim::ParticipatoryProcessesWithUserRole.class_eval do
   alias_method :process_ids_by_process_user_table, :process_ids
 
   def process_ids
-    ids= [process_ids_by_process_user_table]
+    ids = [process_ids_by_process_user_table]
     if user&.department_admin?
       ids << ::Decidim::ParticipatoryProcess
-        .where('decidim_area_id' => user.areas.pluck(:id))
+             .where("decidim_area_id" => user.areas.pluck(:id))
     end
 
     ::Decidim::ParticipatoryProcess.where(id: ids.flatten.uniq)
