@@ -1,2 +1,4 @@
-Decidim.user_roles << 'department_admin'
+# frozen_string_literal: true
+
+Decidim.user_roles << "department_admin"
 Decidim.user_roles.uniq!
