@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require 'decidim/department_admin/admin'
-require 'decidim/department_admin/engine'
-require 'decidim/department_admin/admin_engine'
+require "decidim/department_admin/admin"
+require "decidim/department_admin/engine"
+require "decidim/department_admin/admin_engine"
 
 module Decidim
   # This namespace holds the logic of the `DepartmentAdmin` module. This module
