@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-require 'rails'
-require 'decidim/core'
+require "rails"
+require "decidim/core"
 
 module Decidim
   module DepartmentAdmin
@@ -11,7 +11,7 @@ module Decidim
 
       # make decorators autoload in development env
       config.autoload_paths << File.join(
-        Decidim::DepartmentAdmin::Engine.root, 'app', 'decorators', '{**}'
+        Decidim::DepartmentAdmin::Engine.root, "app", "decorators", "{**}"
       )
 
       routes do
@@ -20,16 +20,16 @@ module Decidim
         # root to: "department_admin#index"
       end
 
-      initializer 'decidim_department_admin.assets' do |app|
-        app.config.assets.precompile += %w[decidim_department_admin_manifest.js decidim_department_admin_manifest.css]
+      initializer "decidim_department_admin.assets" do |app|
+        app.config.assets.precompile += %w(decidim_department_admin_manifest.js decidim_department_admin_manifest.css)
       end
 
-      initializer 'department_admin.permissions_registry' do
+      initializer "department_admin.permissions_registry" do
         # **
         # Modify decidim-admin permissions registry
         # **
-        artifact= ::Decidim::Admin::ApplicationController
-        AdminApplicationControllerPermissions= Class.new(::Decidim::DepartmentAdmin::Permissions)
+        artifact = ::Decidim::Admin::ApplicationController
+        AdminApplicationControllerPermissions = Class.new(::Decidim::DepartmentAdmin::Permissions)
         register_new_permissions_for(artifact, AdminApplicationControllerPermissions)
 
         # **
@@ -38,13 +38,13 @@ module Decidim
 
         # force the concern to be included so that registry is initialized
         # we choose some random class already including it
-        require 'decidim/participatory_processes/admin/categories_controller'
-        artifact= ::Decidim::ParticipatoryProcesses::Admin::Concerns::ParticipatoryProcessAdmin
-        ParticipatoryProcessesAdminConcernPermissions= Class.new(::Decidim::DepartmentAdmin::Permissions)
+        require "decidim/participatory_processes/admin/categories_controller"
+        artifact = ::Decidim::ParticipatoryProcesses::Admin::Concerns::ParticipatoryProcessAdmin
+        ParticipatoryProcessesAdminConcernPermissions = Class.new(::Decidim::DepartmentAdmin::Permissions)
         register_new_permissions_for(artifact, ParticipatoryProcessesAdminConcernPermissions)
 
-        artifact= Decidim::ParticipatoryProcesses::Admin::ApplicationController
-        ParticipatoryProcessesAdminApplicationControllerPermissions= Class.new(::Decidim::DepartmentAdmin::Permissions)
+        artifact = Decidim::ParticipatoryProcesses::Admin::ApplicationController
+        ParticipatoryProcessesAdminApplicationControllerPermissions = Class.new(::Decidim::DepartmentAdmin::Permissions)
         register_new_permissions_for(artifact, ParticipatoryProcessesAdminApplicationControllerPermissions)
 
         # public views produce problems with Devise's current_user
@@ -56,32 +56,32 @@ module Decidim
         # **
         # Modify decidim-assemblies permissions registry
         # **
-        require 'decidim/assemblies/admin/assembly_copies_controller'
-        require 'decidim/assemblies/admin/components_controller'
-        artifact= ::Decidim::Assemblies::Admin::Concerns::AssemblyAdmin
-        AssembliesAdminConcernPermissions= Class.new(::Decidim::DepartmentAdmin::Permissions)
+        require "decidim/assemblies/admin/assembly_copies_controller"
+        require "decidim/assemblies/admin/components_controller"
+        artifact = ::Decidim::Assemblies::Admin::Concerns::AssemblyAdmin
+        AssembliesAdminConcernPermissions = Class.new(::Decidim::DepartmentAdmin::Permissions)
         register_new_permissions_for(artifact, AssembliesAdminConcernPermissions)
 
-        artifact= Decidim::Assemblies::Admin::ApplicationController
-        AssembliesAdminApplicationControllerPermissions= Class.new(::Decidim::DepartmentAdmin::Permissions)
+        artifact = Decidim::Assemblies::Admin::ApplicationController
+        AssembliesAdminApplicationControllerPermissions = Class.new(::Decidim::DepartmentAdmin::Permissions)
         register_new_permissions_for(artifact, AssembliesAdminApplicationControllerPermissions)
       end
 
       # make decorators available to applications that use this Engine
       config.to_prepare do
-        Dir.glob(Decidim::DepartmentAdmin::Engine.root + 'app/decorators/**/*_decorator*.rb').each do |c|
+        Dir.glob(Decidim::DepartmentAdmin::Engine.root + "app/decorators/**/*_decorator*.rb").each do |c|
           require_dependency(c)
         end
       end
 
       config.after_initialize do
-        require 'decidim/participatory_processes/participatory_space'
+        require "decidim/participatory_processes/participatory_space"
         # override participatory_processes space manifest permissions with DepartmentAdmin's one
-        manifest= Decidim.find_participatory_space_manifest(:participatory_processes)
-        manifest.permissions_class_name= 'Decidim::ParticipatoryProcesses::ParticipatorySpacePermissions'
+        manifest = Decidim.find_participatory_space_manifest(:participatory_processes)
+        manifest.permissions_class_name = "Decidim::ParticipatoryProcesses::ParticipatorySpacePermissions"
         # override assemblies space manifest permissions with DepartmentAdmin's one
-        manifest= Decidim.find_participatory_space_manifest(:assemblies)
-        manifest.permissions_class_name= 'Decidim::Assemblies::ParticipatorySpacePermissions'
+        manifest = Decidim.find_participatory_space_manifest(:assemblies)
+        manifest.permissions_class_name = "Decidim::Assemblies::ParticipatorySpacePermissions"
       end
 
       #------------------------------------------------------
@@ -93,9 +93,9 @@ module Decidim
       # Modifies the permissions registry for the given +artifact+ with +new_permissions_class+.
       # NOTE: Previous permissions are cleared and setted to the +new_permissions_class+ for delegation.
       def register_new_permissions_for(artifact, new_permissions_class)
-        chain= ::Decidim.permissions_registry.chain_for(artifact)
+        chain = ::Decidim.permissions_registry.chain_for(artifact)
         # configure old chain, in case new permissions has to delegate
-        new_permissions_class.delegate_chain= chain.dup
+        new_permissions_class.delegate_chain = chain.dup
         # re-set the permissions classes to DepartmentAdmin::Permissions' subclass
         chain.clear
         chain << new_permissions_class
