@@ -1,19 +1,19 @@
 # frozen_string_literal: true
 
-require 'spec_helper'
+require "spec_helper"
 
 module Decidim
   describe InviteUser do
-    describe 'call' do
+    describe "call" do
       let(:organization) { create(:organization) }
 
       let(:current_user) { create(:admin, :confirmed, organization: organization) }
 
       let(:form_params) do
         {
-          name: 'user_name',
-          email: 'user_name@example.org',
-          organization: organization
+          name: "user_name",
+          email: "user_name@example.org",
+          organization: organization,
         }
       end
 
@@ -28,11 +28,11 @@ module Decidim
 
       let(:command) { described_class.new(form, current_user) }
 
-      shared_examples_for 'inviting_users' do
-        context 'when everything is ok' do
-          it 'invites the user with no area' do
+      shared_examples_for "inviting_users" do
+        context "when everything is ok" do
+          it "invites the user with no area" do
             invite_user
-            user = Decidim::User.find_by(email: 'user_name@example.org')
+            user = Decidim::User.find_by(email: "user_name@example.org")
             expect(user).to exist
           end
         end
