@@ -1,3 +1,3 @@
 # frozen_string_literal: true
 
-require 'decidim/department_admin/test/factories'
+require "decidim/department_admin/test/factories"
