@@ -1,15 +1,15 @@
 # frozen_string_literal: true
 
-require 'decidim/dev'
+require "decidim/dev"
 
-ENV['ENGINE_ROOT'] = File.dirname(__dir__)
+ENV["ENGINE_ROOT"] = File.dirname(__dir__)
 
 # require "decidim/admin"
 # require "decidim/core"
 # require "decidim/core/test"
 
-Decidim::Dev.dummy_app_path = File.expand_path(File.join('spec', 'decidim_dummy_app'))
+Decidim::Dev.dummy_app_path = File.expand_path(File.join("spec", "decidim_dummy_app"))
 
-require 'decidim/dev/test/base_spec_helper'
+require "decidim/dev/test/base_spec_helper"
 
-require 'decidim/department_admin'
+require "decidim/department_admin"
