@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require 'spec_helper'
+require "spec_helper"
 
-describe 'Admin explores processes', type: :system do
+describe "Admin explores processes", type: :system do
   let(:organization) { create(:organization) }
   let!(:area) { create(:area, organization: organization) }
   let!(:process) { create(:participatory_process, organization: organization, area: area) }
@@ -16,8 +16,8 @@ describe 'Admin explores processes', type: :system do
 
   describe "when there are admins of all types" do
     let!(:department_admin) do
-      user= create(:user, :confirmed, organization: organization)
-      user.roles << 'department_admin'
+      user = create(:user, :confirmed, organization: organization)
+      user.roles << "department_admin"
       user.areas << area
       user.save!
       user
@@ -30,21 +30,22 @@ describe 'Admin explores processes', type: :system do
 
       it "renders a new column for the process department (aka area)" do
         check_column_header_exists(position: 2, content: "DEPARTMENT/AREA")
-        check_column_data_exists(position: 2, content: area.name['en'])
+        check_column_data_exists(position: 2, content: area.name["en"])
         check_column_data_exists(position: 3, content: department_admin.name)
       end
+
       it "renders a new column for the department admins in the process"
       it "renders a new column for the process admins in the process"
     end
   end
 
-  def check_column_header_exists(position:1, content:'')
+  def check_column_header_exists(position: 1, content: "")
     within "#processes .card-section .table-scroll table thead th:nth-child(#{position})" do
       expect(page).to have_content(content)
     end
   end
 
-  def check_column_data_exists(position:1, content:'')
+  def check_column_data_exists(position: 1, content: "")
     within "#processes .card-section .table-scroll table tbody td:nth-child(#{position})" do
       expect(page).to have_content(content)
     end
