@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require 'spec_helper'
+require "spec_helper"
 
-describe 'Admin manages participatory processes', versioning: true, type: :system do
+describe "Admin manages participatory processes", versioning: true, type: :system do
   let(:organization) { create(:organization) }
   let(:area) { create(:area) }
   let(:department_admin) { create(:department_admin, :confirmed, organization: organization, area: area) }
@@ -20,20 +20,20 @@ describe 'Admin manages participatory processes', versioning: true, type: :syste
     visit decidim_admin_participatory_processes.participatory_processes_path
   end
 
-  it 'should see the import button' do
+  it "sees the import button" do
     visit_admin_processes_list
     expect(page).to have_content("Import")
   end
 
-  it 'should see the export button' do
+  it "sees the export button" do
     visit_admin_processes_list
     expect(page).to have_css("icon--data-transfer-download")
   end
 
-  it 'should see only processes in the same area' do
+  it "sees only processes in the same area" do
     visit_admin_processes_list
-    expect(page).to have_content(participatory_process_w_area.title['en'])
-    expect(page).to_not have_content(participatory_process_wo_area.title['en'])
+    expect(page).to have_content(participatory_process_w_area.title["en"])
+    expect(page).not_to have_content(participatory_process_wo_area.title["en"])
   end
 
   context "when department_admin has a user_role in a participatory_process_wo_area" do
@@ -41,10 +41,10 @@ describe 'Admin manages participatory processes', versioning: true, type: :syste
       create(:participatory_process_user_role, user: department_admin, participatory_process: participatory_process_wo_area)
     end
 
-    it "should see both processes" do
+    it "sees both processes" do
       visit_admin_processes_list
-      expect(page).to have_content(participatory_process_w_area.title['en'])
-      expect(page).to have_content(participatory_process_wo_area.title['en'])
+      expect(page).to have_content(participatory_process_w_area.title["en"])
+      expect(page).to have_content(participatory_process_wo_area.title["en"])
     end
   end
 end
